refactor package name for analytics module
