+++ app/src/main/java/com/kafka/user/playback/PlayerLogger.kt
@@ -4,8 +4,8 @@ package com.kafka.user.playback
 
 import com.sarahang.playback.core.apis.Logger
 import com.sarahang.playback.core.apis.PlayerEventLogger
-import org.kafka.play.logger.Analytics
-import org.kafka.play.logger.EventInfo
+import org.kafka.analytics.logger.Analytics
+import org.kafka.analytics.logger.EventInfo
 import javax.inject.Inject
 
 class KafkaPlayerEventLogger @Inject constructor(
