+++ domain/src/main/java/org/kafka/domain/interactors/account/SignInUser.kt
@@ -2,7 +2,7 @@ package org.kafka.domain.interactors.account
 
 import com.kafka.data.feature.auth.AccountRepository
 import kotlinx.coroutines.withContext
-import org.kafka.play.logger.Analytics
+import org.kafka.analytics.logger.Analytics
 import org.kafka.base.CoroutineDispatchers
 import org.kafka.base.domain.Interactor
 import javax.inject.Inject
