+++ ui/homepage/src/main/java/org/kafka/homepage/recent/RecentViewModel.kt
@@ -10,7 +10,7 @@ import kotlinx.collections.immutable.toPersistentList
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.launch
-import org.kafka.play.logger.Analytics
+import org.kafka.analytics.logger.Analytics
 import org.kafka.base.extensions.stateInDefault
 import org.kafka.domain.interactors.recent.RemoveAllRecentItems
 import org.kafka.domain.observers.ObserveRecentItems
